Move debugWriter declaration to GPUReconstructionCPU.h
