+++ GPU/GPUTracking/Base/GPUReconstruction.h
@@ -424,17 +424,6 @@ class GPUReconstruction
   void debugInit();
   void debugExit();
 
-  struct debugWriter {
-    debugWriter(std::string filenameCSV, bool markdown, uint32_t statNEvents);
-    void header();
-    void row(char type, uint32_t count, std::string name, double gpu_time, double cpu_time, double total_time, std::size_t memSize, std::string nEventReport = "");
-
-   private:
-    std::ofstream streamCSV;
-    bool mMarkdown;
-    uint32_t mStatNEvents;
-  };
-
   static GPUReconstruction* GPUReconstruction_Create_CPU(const GPUSettingsDeviceBackend& cfg);
 };
