+++ GPU/GPUTracking/Base/GPUReconstructionCPU.h
@@ -16,7 +16,9 @@
 #define GPURECONSTRUCTIONICPU_H
 
 #include "GPUReconstructionProcessing.h"
+#include <fstream>
 #include <stdexcept>
+#include <string>
 #include <vector>
 
 namespace Ort
@@ -100,6 +102,17 @@ class GPUReconstructionCPU : public GPUReconstructionProcessing::KernelInterface
   size_t TransferMemoryResourcesHelper(GPUProcessor* proc, int32_t stream, bool all, bool toGPU);
   template <class S, int32_t I = 0, typename... Args>
   void runKernelInterface(krnlSetup&& setup, Args const&... args);
+
+  struct debugWriter {
+    debugWriter(std::string filenameCSV, bool markdown, uint32_t statNEvents);
+    void header();
+    void row(char type, uint32_t count, std::string name, double gpu_time, double cpu_time, double total_time, std::size_t memSize, std::string nEventReport = "");
+
+   private:
+    std::ofstream streamCSV;
+    bool mMarkdown;
+    uint32_t mStatNEvents;
+  };
 };
 
 } // namespace o2::gpu
